+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Version counting is based on semantic versioning (Major.Feature.Patch)
 
 ### YACReader
 * Improve save current page. It includes the comic file name in the default image name and the page number. Now it uses the original image in the comic for image based variants.
+* Fix magnifying glass not using the source material.
 
 ## 10.0.0